Remove connection polling in ctxRW
Removes the 100ms connection polling done by ctxRW in favor of a
simpler approach that sets the read/write deadline when the context
is cancelled.

The 100ms number was effectively the worst-case delay between the
time the context was cancelled and the time that the wait() func
returned by Loopback() could return. For some use cases, the 100ms
delay may not be desired.